+++ src/ast/decode.cr
@@ -5,9 +5,9 @@ module Mint
 
       def initialize(@expression : Expression?,
                      @file : Parser::File,
-                     @from : Int32,
+                     @from : Int64,
                      @type : Type,
-                     @to : Int32)
+                     @to : Int64)
       end
     end
   end
