+++ src/ast/test.cr
@@ -6,8 +6,8 @@ module Mint
       def initialize(@name : StringLiteral,
                      @expression : Block,
                      @file : Parser::File,
-                     @from : Int32,
-                     @to : Int32)
+                     @from : Int64,
+                     @to : Int64)
       end
     end
   end
