+++ src/ast/use.cr
@@ -7,8 +7,8 @@ module Mint
                      @provider : TypeId,
                      @data : Record,
                      @file : Parser::File,
-                     @from : Int32,
-                     @to : Int32)
+                     @from : Int64,
+                     @to : Int64)
       end
     end
   end
